+++ lib/utils/support.ts
@@ -1,0 +1,20 @@
+/**
+ * @private
+ */
+ export function isBrowserSupported() {
+  return !!window.OffscreenCanvas && !(/Mobi/i.test(window.navigator.userAgent)) && !!window.chrome;
+}
+
+/**
+ * Check if the current browser is officially supported by twilio-video-procesors.js.
+ * This is set to `true` for chromium-based desktop browsers.
+ * @example
+ * ```ts
+ * import { isSupported } from '@twilio/video-processors-sdk';
+ *
+ * if (isSupported) {
+ *   // Initialize the background processors
+ * }
+ * ```
+ */
+ export const isSupported = isBrowserSupported();