+++ tests/unit/index.ts
@@ -2,6 +2,11 @@ import { Canvas } from './mocks/Canvas';
 
 const root = global as any;
 root.OffscreenCanvas = root.OffscreenCanvas || Canvas;
+root.window = root.window || {
+  navigator : {
+    userAgent : 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.77 Safari/537.36'
+  }
+}
 root.document = root.document || {
   createElement(name: string) {
     if (name === 'canvas') {
@@ -10,8 +15,9 @@ root.document = root.document || {
   }
 };
 
-import './utils/Benchmark';
-import './processors/grayscale';
 import './processors/background/BackgroundProcessor';
 import './processors/background/GaussianBlurBackgroundProcessor';
 import './processors/background/VirtualBackgroundProcessor';
+import './processors/grayscale';
+import './utils/Benchmark';
+import './utils/support';